Add increment and decrement operators.
These are really bad, codegen-wise, because there are two competing
optimization opportunities and everyone seems to pick the wrong one.
The correct optimization is to notice that "add 1, test for less-than-1,
add with carry" can be turned into "ADD/ADC". The incorrect optimization
is to notice that "add 1, test for less-than-1" can be turned into
"ADD/SETE", which then doesn't chain into "ADD/ADC"
the way we want it to.

https://bugs.llvm.org/show_bug.cgi?id=31754

The other bad thing going on is that `*p = (*q)++` generates really
bad code thanks to the possible aliasing between `*p` and `*q`.
https://godbolt.org/z/kaJtM5